refactor(library): 패키지 구조 변경 및 수행한 기능목록 체크
