+++ README.md
@@ -31,15 +31,11 @@
         - [ ] 이미 돌이 있는 자리에 놓을 수 없다.
         - [ ] 검정 돌을 놓는다.
         - [ ] 흰 돌을 놓는다.
-    -[x] 차례를 반환한다.
-- Rule (interface)
-    - 승리 조건
-        -[ ] 5이상이 되면 승리
-    - 흑돌 금지 조건
-        -[ ] 3-3
-        -[ ] 4-4
-        -[ ] 4-3-3
-        -[ ] 4-4-3
-        -[ ] 장목
-        -[ ] 예외: 위 사항들에 해당되면서도 5가 만들어지면 승리
+    -[ ] 차례를 반환한다.
+- Rule (library)
+    - 조건
+        -[x] 3-3
+        -[x] 4-4
+        -[x] 장목
+        -[x] 오목
