+++ src/test/kotlin/library/exceedfive/ExceedFiveDummy.kt
@@ -1,20 +1,22 @@
-package library
+package library.exceedfive
 
 import domain.Coordinate
 import domain.CoordinateState
 import domain.Position
 
-object ExactlyFiveDummy {
+object ExceedFiveDummy {
+
     // 3
-    // 4 X X X ? X
-    //   3 4 5 6 7
-    fun getExactlyFiveCoordinate() = Position(Coordinate(4), Coordinate(6))
-    fun getExactlyFiveBoard(coordinateState: CoordinateState): List<MutableList<CoordinateState>> {
+    // 4 X X X ? X X
+    //   3 4 5 6 7 8
+    fun getExceedFiveCoordinate() = Position(Coordinate(4), Coordinate(6))
+    fun getExceedFiveBoard(coordinateState: CoordinateState): List<MutableList<CoordinateState>> {
         val board = List(15) { MutableList(15) { CoordinateState.EMPTY } }
         board[4][3] = coordinateState
         board[4][4] = coordinateState
         board[4][5] = coordinateState
         board[4][7] = coordinateState
+        board[4][8] = coordinateState
 
         return board
     }
